Project 2 - Bill Extractor AI Assistant/requirements.txt
streamlit
langchain
langchain-google-genai
pypdf
pandas
python-dotenv
